Increase the logic of the feed in the api layer

+++ cmd/api/handlers/handler.go
@@ -18,7 +18,7 @@ func SendResponse(c *gin.Context, err error, data interface{}) {
 }
 
 type Response struct {
-	Code    int64       `json:"code"`
+	Code    int32       `json:"code"`
 	Message string      `json:"message"`
 	Data    interface{} `json:"data"`
 }
